requirements.txt
alembic==1.6.5
attrs==21.2.0
Flask-Migrate==2.7.0
flask-restx==0.5.0
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.1
SQLAlchemy==1.4.23
